feat: create hooks
